test(Station, Line): 공통 설정 분리
